more elegant handling of TBD team
